requirements.txt
PySide6>=6.0.0
Pillow>=9.0.0
orjson>=3.8.0
